Added editing and deleting a game
Added func to edit and delete your game when you are the creator. Fixes

+++ Mist2/games/models.py
@@ -29,8 +29,8 @@ class Game(models.Model):
     user = models.ForeignKey(to=UserModel, on_delete=models.CASCADE)
     game_photo = models.ImageField(blank=True, null=True, validators=(image_size_validator_5mb,),
                                    upload_to='images//')
-    size = models.PositiveIntegerField(null=False, blank=False)
-    price = models.PositiveIntegerField(null=False, blank=False)
+    size = models.DecimalField(max_digits=10, decimal_places=2, null=False, blank=False)
+    price = models.DecimalField(max_digits=10, decimal_places=2, null=False, blank=False)
     description = models.TextField(max_length=300, validators=(MinLengthValidator(10),), blank=True, null=True)
 
     def __str__(self):
